SAK-22108 fixed a bug related with Oracle deployment. The lti tool id is a BigDecimal field, and cannot cast to Integer.
git-svn-id: https://source.sakaiproject.org/svn/site-manage/trunk@111003 66ffb92e-73f9-0310-93c1-f5514f145a0a

+++ site-manage/site-manage-tool/tool/src/java/org/sakaiproject/site/tool/SiteAction.java
@@ -1589,7 +1589,7 @@ public class SiteAction extends PagedResourceActionII {
 				List<Map<String,Object>> contents = m_ltiService.getContents(null,null,0,0);
 				HashMap<String, Map<String, Object>> linkedLtiContents = new HashMap<String, Map<String, Object>>();
 				for ( Map<String,Object> content : contents ) {
-					String ltiToolId = ((Integer) content.get(m_ltiService.LTI_TOOL_ID)).toString();
+					String ltiToolId = content.get(m_ltiService.LTI_TOOL_ID).toString();
 					String siteId = StringUtils.trimToNull((String) content.get(m_ltiService.LTI_SITE_ID));
 					if (siteId != null)
 					{
